Move PagesSerde and PagesSerdeUtil to SPI
