+++ presto-main/src/test/java/com/facebook/presto/execution/buffer/TestPagesSerde.java
@@ -16,6 +16,7 @@ package com.facebook.presto.execution.buffer;
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.block.BlockBuilder;
+import com.facebook.presto.spi.page.PagesSerde;
 import com.facebook.presto.spi.type.Type;
 import com.google.common.collect.ImmutableList;
 import io.airlift.slice.DynamicSliceOutput;
@@ -25,9 +26,9 @@ import org.testng.annotations.Test;
 import java.util.Iterator;
 import java.util.List;
 
-import static com.facebook.presto.execution.buffer.PagesSerdeUtil.readPages;
-import static com.facebook.presto.execution.buffer.PagesSerdeUtil.writePages;
 import static com.facebook.presto.operator.PageAssertions.assertPageEquals;
+import static com.facebook.presto.spi.page.PagesSerdeUtil.readPages;
+import static com.facebook.presto.spi.page.PagesSerdeUtil.writePages;
 import static com.facebook.presto.spi.type.BigintType.BIGINT;
 import static com.facebook.presto.spi.type.VarcharType.VARCHAR;
 import static org.testng.Assert.assertEquals;
