+++ presto-main/src/test/java/com/facebook/presto/operator/MockExchangeRequestProcessor.java
@@ -20,9 +20,9 @@ import com.facebook.airlift.http.client.testing.TestingHttpClient;
 import com.facebook.airlift.http.client.testing.TestingResponse;
 import com.facebook.presto.client.PrestoHeaders;
 import com.facebook.presto.execution.buffer.BufferResult;
-import com.facebook.presto.execution.buffer.PagesSerde;
-import com.facebook.presto.execution.buffer.PagesSerdeUtil;
 import com.facebook.presto.spi.Page;
+import com.facebook.presto.spi.page.PagesSerde;
+import com.facebook.presto.spi.page.PagesSerdeUtil;
 import com.facebook.presto.spi.page.SerializedPage;
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
