+++ presto-main/src/test/java/com/facebook/presto/operator/TestExchangeClient.java
@@ -18,9 +18,9 @@ import com.facebook.airlift.http.client.Response;
 import com.facebook.airlift.http.client.testing.TestingHttpClient;
 import com.facebook.presto.block.BlockAssertions;
 import com.facebook.presto.execution.TaskId;
-import com.facebook.presto.execution.buffer.PagesSerde;
 import com.facebook.presto.memory.context.SimpleLocalMemoryContext;
 import com.facebook.presto.spi.Page;
+import com.facebook.presto.spi.page.PagesSerde;
 import com.facebook.presto.spi.page.SerializedPage;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.util.concurrent.Futures;
